test: fix tics abstraction interpretations
