+++ src/features/activities/components/ActivityDetails/components/ActivityDetailsButtons/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./ActivityDetailsButtons";
